rename fib to fip
