+++ src/aind_metadata_mapper/fip/__init__.py
