+++ tests/resources/fip/000000_ophys_session.json
