+++ tests/resources/fip/example_from_teensy.txt
