--- tests/test_fib/__init__.py
@@ -1,1 +1,0 @@
-"""Tests fib package"""
