+++ tests/test_fip/__init__.py
@@ -1,0 +1,1 @@
+"""Tests fip package"""
